ci: fail loudly when DB-backed tests cannot run
