+++ .github/workflows/build-and-notarize.yml
@@ -34,6 +34,8 @@ jobs:
         env:
           # keeps require("electron") from throwing when postinstall was skipped
           ELECTRON_OVERRIDE_DIST_PATH: /tmp
+          # a skipped DB test here means the rebuild step above is gone: fail instead
+          REQUIRE_DB_TESTS: "1"
 
   build-linux:
     needs: tests
