+++ test/helpers/snippetsDatabase.test.js
@@ -24,13 +24,7 @@ Module._load = function patchedLoad(request, parent, isMain) {
 process.env.NODE_ENV = "test";
 
 const DatabaseManager = require("../../src/helpers/database.js");
-
-function isNativeBindingUnavailable(error) {
-  const message = String(error?.message || error);
-  return (
-    message.includes("NODE_MODULE_VERSION") || message.includes("Could not locate the bindings file")
-  );
-}
+const { skipOrFail } = require("./harness/db.js");
 
 function createDb(t) {
   userDataDir = fs.mkdtempSync(path.join(os.tmpdir(), "openwhispr-snippets-db-"));
@@ -40,21 +34,15 @@ function createDb(t) {
     probe.close();
     fs.rmSync(path.join(userDataDir, "probe.db"), { force: true });
   } catch (error) {
-    if (isNativeBindingUnavailable(error)) {
-      t.skip("better-sqlite3 native binding is not available for this Node runtime");
-      return null;
-    }
-    throw error;
+    skipOrFail(t, error);
+    return null;
   }
 
   try {
     return new DatabaseManager();
   } catch (error) {
-    if (isNativeBindingUnavailable(error)) {
-      t.skip("better-sqlite3 native binding is not available for this Node runtime");
-      return null;
-    }
-    throw error;
+    skipOrFail(t, error);
+    return null;
   }
 }
