+++ test/helpers/translationHistory.test.js
@@ -24,18 +24,7 @@ Module._load = function patchedLoad(request, parent, isMain) {
 process.env.NODE_ENV = "test";
 
 const DatabaseManager = require("../../src/helpers/database.js");
-
-// The repo's better-sqlite3 binding is built for Electron's ABI, so a plain `node`
-// runtime cannot dlopen it. Skip cleanly instead of failing in that case.
-function isNativeBindingUnavailable(error) {
-  const message = String(error?.message || error);
-  return (
-    message.includes("NODE_MODULE_VERSION") ||
-    message.includes("Could not locate the bindings file") ||
-    message.includes("ERR_DLOPEN_FAILED") ||
-    error?.code === "ERR_DLOPEN_FAILED"
-  );
-}
+const { skipOrFail } = require("./harness/db.js");
 
 function createDb(t) {
   userDataDir = fs.mkdtempSync(path.join(os.tmpdir(), "openwhispr-translation-db-"));
@@ -45,21 +34,15 @@ function createDb(t) {
     probe.close();
     fs.rmSync(path.join(userDataDir, "probe.db"), { force: true });
   } catch (error) {
-    if (isNativeBindingUnavailable(error)) {
-      t.skip("better-sqlite3 native binding is not available for this Node runtime");
-      return null;
-    }
-    throw error;
+    skipOrFail(t, error);
+    return null;
   }
 
   try {
     return new DatabaseManager();
   } catch (error) {
-    if (isNativeBindingUnavailable(error)) {
-      t.skip("better-sqlite3 native binding is not available for this Node runtime");
-      return null;
-    }
-    throw error;
+    skipOrFail(t, error);
+    return null;
   }
 }
